+++ core/src/main/java/io/snabble/sdk/checkout/CheckoutApi.kt
@@ -4,15 +4,14 @@ import com.google.gson.Gson
 import com.google.gson.JsonObject
 import com.google.gson.annotations.SerializedName
 import com.google.gson.reflect.TypeToken
-import io.snabble.sdk.payment.PaymentCredentials
-import io.snabble.sdk.Product
-import io.snabble.sdk.coupons.Coupon
 import io.snabble.sdk.FulfillmentState
 import io.snabble.sdk.PaymentMethod
+import io.snabble.sdk.Product
+import io.snabble.sdk.coupons.Coupon
+import io.snabble.sdk.payment.PaymentCredentials
 import io.snabble.sdk.shoppingcart.data.cart.BackendCart
 import java.io.Serializable
-import java.lang.Exception
-import java.util.*
+import java.util.Date
 
 /**
  * Interface for the snabble Checkout API
@@ -91,7 +90,6 @@ interface CheckoutInfoResult {
     fun onNoShopFound()
     fun onInvalidProducts(products: List<Product>)
     fun onNoAvailablePaymentMethodFound()
-    fun onInvalidDepositReturnVoucher()
     fun onUnknownError()
     fun onConnectionError()
 }
@@ -215,10 +213,27 @@ data class CheckoutInfo(
 )
 
 data class Violation(
-    val type: String? = null,
+    val type: ViolationType? = null,
     val refersTo: String? = null,
     val message: String? = null,
 )
+
+enum class ViolationType {
+    @SerializedName("deposit_return_voucher_already_redeemed")
+    DEPOSIT_RETURN_ALREADY_REDEEMED,
+
+    @SerializedName("deposit_return_voucher_duplicate")
+    DEPOSIT_RETURN_DUPLICATED,
+
+    @SerializedName("coupon_already_voided")
+    COUPON_ALREADY_VOIDED,
+
+    @SerializedName("coupon_currently_not_valid")
+    COUPON_CURRENTLY_NOT_VALID,
+
+    @SerializedName("coupon_invalid")
+    COUPON_INVALID,
+}
 
 data class LineItem(
     val id: String? = null,