+++ core/src/main/java/io/snabble/sdk/shoppingcart/ShoppingCartUpdater.kt
@@ -98,11 +98,6 @@ internal class ShoppingCartUpdater(
                     error(requestSucceeded = true)
                 }
 
-                override fun onInvalidDepositReturnVoucher() {
-                    cart.setInvalidDepositReturnVoucher(true)
-                    error(requestSucceeded = true)
-                }
-
                 override fun onUnknownError() {
 
                     error(requestSucceeded = false)
@@ -168,8 +163,6 @@ internal class ShoppingCartUpdater(
             invalidProducts = null
             checkLimits()
             notifyPriceUpdate(this)
-            forEach {
-            }
         }
     }
 
@@ -290,10 +283,10 @@ internal class ShoppingCartUpdater(
         else -> false
     }
 
-    private fun containsReturnDepositReturnVouchers(signedCheckoutInfo: SignedCheckoutInfo): Boolean {
-        val (_, lineItems) = deserializedCheckoutInfo(signedCheckoutInfo) ?: return false
-        return lineItems.any { it.type == LineItemType.DEPOSIT_RETURN_VOUCHER }
-    }
+    private fun containsReturnDepositReturnVouchers(signedCheckoutInfo: SignedCheckoutInfo): Boolean =
+        deserializedCheckoutInfo(signedCheckoutInfo)?.lineItems
+            ?.any { it.type == LineItemType.DEPOSIT_RETURN_VOUCHER }
+            ?: false
 
     private fun getToBeReplacedSkus(signedCheckoutInfo: SignedCheckoutInfo): List<String?> {
         val (_, lineItems) = deserializedCheckoutInfo(signedCheckoutInfo) ?: return emptyList()