+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 ### Changed
 ### Removed
+* ui/core: Remove everything related to the old deposit return voucher feature
 ### Fixed
 
 ## [0.79.0]